히스토리 추가 Todo: localDateTime 직렬화

+++ src/main/java/com/ohgiraffers/r_pakabe/flow/sceneHistory/command/domain/model/SceneHistory.java
@@ -25,7 +25,7 @@ public class SceneHistory {
     @Id
     private String id;
 
-    private Long roomNum;
+    private Integer roomNum;
 
     private String history;
